+++ examples/workflows/cockroach_folder/main.tf
@@ -20,10 +20,10 @@ variable "cluster_name" {
   nullable = false
 }
 
-variable "serverless_spend_limit" {
+variable "request_unit_rate_limit" {
   type     = number
   nullable = false
-  default  = 0
+  default  = 1000
 }
 
 variable "cloud_provider" {
@@ -89,7 +89,9 @@ resource "cockroach_cluster" "example" {
   name           = var.cluster_name
   cloud_provider = var.cloud_provider
   serverless = {
-    spend_limit = var.serverless_spend_limit
+    usage_limits = {
+      request_unit_rate_limit = var.request_unit_rate_limit
+    }
   }
   regions   = [for r in var.cloud_provider_regions : { name = r }]
   parent_id = cockroach_folder.example-folder-parent.id