+++ internal/provider/connection_string_data_source_test.go
@@ -59,7 +59,7 @@ func TestIntegrationConnectionStringDataSource(t *testing.T) {
 		CloudProvider: "GCP",
 		State:         "CREATED",
 		Config: client.ClusterConfig{
-			Serverless: &client.ServerlessClusterConfig{},
+			Shared: &client.SharedClusterConfig{},
 		},
 		Regions: []client.Region{
 			{
@@ -137,7 +137,7 @@ func getTestConnectionStringResourceConfig(name, sqlPassword string) string {
 resource "cockroach_cluster" "test" {
 	name           = "%s"
 	cloud_provider = "GCP"
-	serverless = {}
+	shared = {}
 	regions = [{
 		name = "us-central1"
 	}]