+++ internal/provider/database_resource_test.go
@@ -59,14 +59,14 @@ func TestIntegrationDatabaseResource(t *testing.T) {
 	defer HookGlobal(&NewService, func(c *client.Client) client.Service {
 		return s
 	})()
-	spendLimit := int32(1)
 	cluster := client.Cluster{
 		Name:          clusterName,
 		Id:            uuid.Nil.String(),
+		Plan:          "BASIC",
 		CloudProvider: "GCP",
 		Config: client.ClusterConfig{
-			Serverless: &client.ServerlessClusterConfig{
-				SpendLimit: &spendLimit,
+			Shared: &client.SharedClusterConfig{
+				RoutingId: "routing-id",
 			},
 		},
 		State: "CREATED",
@@ -121,7 +121,7 @@ func testDatabaseResource(
 	t *testing.T, clusterName, databaseName, newDatabaseName string, useMock bool,
 ) {
 	var (
-		clusterResourceName = "cockroach_cluster.serverless"
+		clusterResourceName = "cockroach_cluster.shared"
 		resourceName        = "cockroach_database.test_database"
 	)
 
@@ -189,20 +189,18 @@ func testDatabaseExists(resourceName, clusterResourceName string) resource.TestC
 
 func getTestDatabaseResourceConfig(clusterName, databaseName string) string {
 	return fmt.Sprintf(`
-resource "cockroach_cluster" "serverless" {
+resource "cockroach_cluster" "shared" {
     name           = "%s"
     cloud_provider = "GCP"
-    serverless = {
-        spend_limit = 1
-    }
-	regions = [{
-		name = "us-central1"
-	}]
+    shared = {}
+    regions = [{
+        name = "us-central1"
+    }]
 }
 
 resource "cockroach_database" "test_database" {
   name = "%s"
-  cluster_id = cockroach_cluster.serverless.id
+  cluster_id = cockroach_cluster.shared.id
 }
 `, clusterName, databaseName)
 }